requirements.txt
streamlit==1.35.0
langchain==0.1.20
pdfplumber==0.10.3
numpy==1.26.4
scikit-learn==1.4.2
requests==2.31.0
python-dotenv==1.0.1
